Move cloud identity android tests
