+++ apps/android/data/local/src/androidTest/java/com/flashcardsopensourceapp/data/local/cloudidentity/CloudGuestSessionCoordinatorTest.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.data.local
+package com.flashcardsopensourceapp.data.local.cloudidentity
 
 import android.content.Context
 import androidx.test.ext.junit.runners.AndroidJUnit4
