+++ apps/android/data/local/src/androidTest/java/com/flashcardsopensourceapp/data/local/cloudidentity/CloudIdentityTestFixtures.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.data.local
+package com.flashcardsopensourceapp.data.local.cloudidentity
 
 import android.content.Context
 import com.flashcardsopensourceapp.data.local.database.SyncStateEntity
